Fix cargo-audit vulnerability and clippy warnings
- Modified `Cargo.toml` to disable default features for `sqlx` and switch from `tls-native-tls` to `tls-rustls`. This successfully removes the vulnerable `rsa` crate pulled in by `sqlx-mysql`.
- Applied automated fixes via `cargo clippy --fix` for simplified iterations (`needless_range_loop`), redundant operations (`double_parens`), and correct parameter types (`ptr_arg`).
- Manually fixed unresolved variables caused by clippy applying `_` to variables still in use (`elapsed_ms`, `query`).
- Addressed mutation complaints by correctly marking `weight` as `mut` in `src/core/vrp/utils.rs`.
- Validated fixes by confirming `cargo audit`, `cargo check`, and `cargo test` pass successfully.

+++ src/core/vrp/solvers/neural_guided.rs
@@ -137,7 +137,7 @@ fn move_features_oropt(
         chain_len as f32,
         if is_rev { 1.0 } else { 0.0 },
         (delta / (remove_gain.abs() + 1e-6)) as f32,
-        (chain_len as f32 / route_len as f32) as f32,
+        (chain_len as f32 / route_len as f32),
         0.0,
         0.0,
         0.0,
